Initial tests for certificate views
Refactors the test suite to use a PostgreSQL database in order to test
the certificate listings before and after a certificate request. This
will be expanded to assert expected contents later on.
It's not clear why the random port OIDC mocking server with Step CA
running inside a container worked before - now it no longer does, and
the stack is changed to use a static port with host networking
(further explanation in test/setup.sh).

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>