Prevent token reuse further
Only one certificate can be issued per token, the previous patch was not
sufficient. Now redirect the user to a fresh login for every certificate
request with an already used token, regardless of whether the token was
really already used for a certificate request and not just for other
activity. This is more reliable and given certificate requests are a
powerful operation anyways, the additional login flow is deemed
reasonable.

Fixes: 56feebd2d44a ("Prevent token reuse")
Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>